+++ frontend/src/public/components/widgets/summary/metricTopButtons.vue
@@ -19,36 +19,6 @@
 								class="close lightButton" title="Zoom al indicador" v-on:click="zoomExtents()">
 					<i class="fas fa-expand-arrows-alt" style="margin-left: 2px;" />
 				</button>
-
-				<span class="dropdown" v-show="Use.UseUrbanity">
-					<button type="button" class="close lightButton" data-toggle="dropdown" title="Urbano/Rural">
-						<i class="fas fa-users" v-text="getUrbanityTextActive()"/>
-					</button>
-					<ul class="dropdown-menu dropdownMargin">
-						<li>
-							<button type="button" class="close lightButton btn-full" v-on:click="changeUrbanity('N')">Todo</button>
-						</li>
-						<li>
-							<button type="button" class="close lightButton btn-full" v-on:click="changeUrbanity('UD')">Urbano</button>
-						</li>
-						<li>
-							<button type="button" class="close lightButton btn-full" v-on:click="changeUrbanity('U')">Urbano Agrupado</button>
-						</li>
-						<li>
-							<button type="button" class="close lightButton btn-full" v-on:click="changeUrbanity('D')">Urbano Disperso</button>
-						</li>
-						<li>
-							<button type="button" class="close lightButton btn-full" v-on:click="changeUrbanity('RL')">Rural</button>
-						</li>
-						<li>
-							<button type="button" class="close lightButton btn-full" v-on:click="changeUrbanity('R')">Rural Agrupado</button>
-						</li>
-						<li>
-							<button type="button" class="close lightButton btn-full" v-on:click="changeUrbanity('L')">Rural Disperso</button>
-						</li>
-					</ul>
-				</span>
-
 			</h5>
 		</div>
 	</div>
@@ -71,7 +41,6 @@ export default {
 	data() {
 		return {
 			work: {},
-			urbanity: '',
 		};
 	},
 	methods: {
@@ -114,29 +83,6 @@ export default {
 			window.SegMap.MapsApi.FitEnvelope(extents, true);
 			this.$refs.zoomExtentsBtn.blur();
 		},
-		getUrbanityTextActive() {
-			if(this.urbanity === 'N') {
-				return '';
-			}else if(this.urbanity === 'UD') {
-				return ' - U';
-			}else if(this.urbanity === 'U') {
-				return ' - UA';
-			}else if(this.urbanity === 'D') {
-				return ' - UD';
-			}else if(this.urbanity === 'RL') {
-				return ' - R';
-			}else if(this.urbanity === 'R') {
-				return ' - RA';
-			}else if(this.urbanity === 'L') {
-				return ' - RD';
-			}
-		},
-		changeUrbanity(mode) {
-			this.metric.properties.SelectedUrbanity = mode;
-			window.SegMap.SaveRoute.UpdateRoute();
-			window.SegMap.UpdateMap();
-			this.urbanity = mode;
-		},
 		shouldClearSelection(intersect, extents) {
 			if (intersect === null) {
 				return true;
@@ -163,21 +109,9 @@ export default {
   font-size: .8em;
   }
 
-.lightButton {
-	font-size: 12px;
-  padding: 4px;
-	line-height: 1em;
-}
-
 .activeButton {
 	opacity: .45;
 }
-.btn-full {
-	width: 100%;
-}
-.dropdownMargin {
-	left: -70px;
-	right: auto;
-	margin-top: 25px;
-}
+
+
 </style>